Move duplicated code into a trait that both processors can use.
